+++ wdevs-tax-switch.php
@@ -15,8 +15,8 @@
  * @wordpress-plugin
  * Plugin Name:          Tax Switch for WooCommerce
  * Plugin URI:           https://wijnberg.dev
- * Description:          Let customers toggle between inclusive and exclusive VAT pricing in your WooCommerce store.
- * Version:              1.7.0
+ * Description:          Let customers toggle between inclusive and exclusive tax pricing in your WooCommerce store.
+ * Version:              1.8.0
  * Author:               Wijnberg Developments
  * Author URI:           https://wijnberg.dev/
  * License:              GPL-2.0+
@@ -26,7 +26,7 @@
  * Requires PHP:         7.2
  * Requires at least:    5.0
  * WC requires at least: 7.0.0
- * WC tested up to:      10.7.0
+ * WC tested up to:      10.9.1
  * Requires Plugins:     woocommerce
  */
 
@@ -40,7 +40,7 @@ if ( ! defined( 'WPINC' ) ) {
  * Start at version 1.0.0 and use SemVer - https://semver.org
  * Rename this for your plugin and update it as you release new versions.
  */
-define( 'WDEVS_TAX_SWITCH_VERSION', '1.7.0' );
+define( 'WDEVS_TAX_SWITCH_VERSION', '1.8.0' );
 
 /**
  * The code that runs during plugin activation.